+++ build.gradle.kts
@@ -1,7 +1,7 @@
 // Top-level build file where you can add configuration options common to all sub-projects/modules.
 plugins {
-    id("com.android.application") version "7.4.2" apply false
-    id("com.android.library") version "7.4.2" apply false
+    id("com.android.application") version "8.0.0" apply false
+    id("com.android.library") version "8.0.0" apply false
     id("com.google.devtools.ksp") version "1.8.10-1.0.9" apply false
     id("org.jetbrains.kotlin.android") version DependencyGradle.KOTLIN_VERSION apply false
     id("org.jetbrains.kotlin.jvm") version DependencyGradle.KOTLIN_VERSION apply false
