+++ frogocoresdk/build.gradle.kts
@@ -10,13 +10,13 @@ group = ProjectSetting.PROJECT_LIB_ID_CORE_SDK
 version = ProjectSetting.PROJECT_VERSION_NAME
 
 java {
-    sourceCompatibility = JavaVersion.VERSION_11
-    targetCompatibility = JavaVersion.VERSION_11
+    sourceCompatibility = JavaVersion.VERSION_17
+    targetCompatibility = JavaVersion.VERSION_17
 }
 
 tasks.withType<KotlinCompile> {
     kotlinOptions {
-        jvmTarget = JavaVersion.VERSION_11.toString()
+        jvmTarget = "17"
     }
 }
