--- frogolog/build.gradle.kts
@@ -1,74 +1,0 @@
-plugins {
-    id("com.android.library")
-    id("org.jetbrains.kotlin.android")
-    id("com.google.devtools.ksp")
-    `maven-publish`
-}
-
-android {
-
-    compileSdk = ProjectSetting.PROJECT_COMPILE_SDK
-    namespace = ProjectSetting.PROJECT_LIB_ID_LOG
-
-    defaultConfig {
-        minSdk = ProjectSetting.PROJECT_MIN_SDK
-        targetSdk = ProjectSetting.PROJECT_TARGET_SDK
-
-        multiDexEnabled = true
-        vectorDrawables.useSupportLibrary = true
-
-        testInstrumentationRunner = "androidx.test.runner.AndroidJUnitRunner"
-
-        consumerProguardFile("consumer-rules.pro")
-    }
-
-    buildTypes {
-        release {
-            isMinifyEnabled = false
-
-            proguardFiles(
-                getDefaultProguardFile("proguard-android-optimize.txt"),
-                "proguard-rules.pro"
-            )
-        }
-    }
-
-    compileOptions {
-        sourceCompatibility = JavaVersion.VERSION_11
-        targetCompatibility = JavaVersion.VERSION_11
-    }
-
-    kotlin {
-        jvmToolchain {
-            languageVersion.set(JavaLanguageVersion.of("11"))
-        }
-    }
-
-}
-
-afterEvaluate {
-    publishing {
-        publications {
-
-            // Creates a Maven publication called "release".
-            register("release", MavenPublication::class) {
-
-                // Applies the component for the release build variant.
-                // NOTE : Delete this line code if you publish Native Java / Kotlin Library
-                from(components["release"])
-
-                // Library Package Name (Example : "com.frogobox.androidfirstlib")
-                // NOTE : Different GroupId For Each Library / Module, So That Each Library Is Not Overwritten
-                groupId = ProjectSetting.PROJECT_LIB_ID_LOG
-
-                // Library Name / Module Name (Example : "androidfirstlib")
-                // NOTE : Different ArtifactId For Each Library / Module, So That Each Library Is Not Overwritten
-                artifactId = ProjectSetting.MODULE_NAME_LOG
-
-                // Version Library Name (Example : "1.0.0")
-                version = ProjectSetting.PROJECT_VERSION_NAME
-
-            }
-        }
-    }
-}
