+++ frogosdk/src/main/java/com/frogobox/sdk/log/FLog.kt
@@ -1,12 +1,12 @@
-package com.frogobox.log
+package com.frogobox.sdk.log
 
 import android.content.Context
 import android.util.Log
 import android.widget.Toast
-import com.frogobox.log.LogConstant.SIMPLE_MESSSAGE
+import com.frogobox.sdk.log.LogConstant.SIMPLE_MESSSAGE
 
 /**
- * Created by Faisal Amir on 14/01/2021
+ * Created by Faisal Amir on 14/02/2021
  * LogCat Source Code
  * -----------------------------------------
  * Name     : Muhammad Faisal Amir
@@ -17,78 +17,72 @@ import com.frogobox.log.LogConstant.SIMPLE_MESSSAGE
  * All rights reserved
  *
  */
-object FrogoLog : IFrogoLog {
+
+object FLog : ILog {
 
     // Tag for get class name this function
     private fun tag(): String {
         return this.javaClass.simpleName
     }
 
-    // Function for get line number of code
-    private fun lineNumber(): String {
-        return Thread.currentThread().stackTrace[4].let {
-            "${it.className.substringAfterLast(".")}.${it.methodName}(${it.fileName}:${it.lineNumber})"
-        }
-    }
-
     // Function Log Simple Debug without message params
     override fun d() {
-        Log.d(tag(), "${lineNumber()}: $SIMPLE_MESSSAGE")
+        Log.d(tag(), SIMPLE_MESSSAGE)
     }
 
     // Function Log Debug
     override fun d(msg: String?) {
-        Log.d(tag(), "${lineNumber()}: $msg")
+        Log.d(tag(), "$msg")
     }
 
     // Function Log Simple Verbose without message params
     override fun v() {
-        Log.v(tag(), "${lineNumber()}: $SIMPLE_MESSSAGE")
+        Log.v(tag(), SIMPLE_MESSSAGE)
     }
 
     // Function Log Verbose
     override fun v(msg: String?) {
-        Log.v(tag(), "${lineNumber()}: $msg")
+        Log.v(tag(), "$msg")
     }
 
     // Function Log Simple Info without message params
     override fun i() {
-        Log.i(tag(), "${lineNumber()}: $SIMPLE_MESSSAGE")
+        Log.i(tag(), SIMPLE_MESSSAGE)
     }
 
     // Function Log Info
     override fun i(msg: String?) {
-        Log.i(tag(), "${lineNumber()}: $msg")
+        Log.i(tag(), "$msg")
     }
 
     // Function Log Simple Warn without message params
     override fun w() {
-        Log.w(tag(), "${lineNumber()}: $SIMPLE_MESSSAGE")
+        Log.w(tag(), SIMPLE_MESSSAGE)
     }
 
     // Function Log Warn
     override fun w(msg: String?) {
-        Log.w(tag(), "${lineNumber()}: $msg")
+        Log.w(tag(), "$msg")
     }
 
     // Function Log Warn
     override fun w(e: Throwable?) {
-        Log.w(tag(), "${lineNumber()}: ${e?.localizedMessage}")
+        Log.w(tag(), "${e?.localizedMessage}")
     }
 
     // Function Log Warn
     override fun w(e: Exception?) {
-        Log.w(tag(), "${lineNumber()}: ${e?.localizedMessage}")
+        Log.w(tag(), "${e?.localizedMessage}")
     }
 
     // Function Log Simple Error without message params
     override fun e() {
-        Log.e(tag(), "${lineNumber()}: $SIMPLE_MESSSAGE")
+        Log.e(tag(), SIMPLE_MESSSAGE)
     }
 
     // Function Log Error
     override fun e(msg: String?) {
-        Log.e(tag(), "${lineNumber()}: $msg")
+        Log.e(tag(), "$msg")
     }
 
     // ---------------------------------------------------------------------------------------------
@@ -101,8 +95,8 @@ object FrogoLog : IFrogoLog {
 
     // Function Log Debug
     override fun d(msg: String?, context: Context) {
-        Log.d(tag(), "${lineNumber()}: $msg")
-        Toast.makeText(context, "${lineNumber()}: $msg", Toast.LENGTH_LONG).show()
+        Log.d(tag(), "$msg")
+        Toast.makeText(context, "$msg", Toast.LENGTH_LONG).show()
     }
 
     // Function Log Simple Verbose without message params
@@ -113,8 +107,8 @@ object FrogoLog : IFrogoLog {
 
     // Function Log Verbose
     override fun v(msg: String?, context: Context) {
-        Log.v(tag(), "${lineNumber()}: $msg")
-        Toast.makeText(context, "${lineNumber()}: $msg", Toast.LENGTH_LONG).show()
+        Log.v(tag(), "$msg")
+        Toast.makeText(context, "$msg", Toast.LENGTH_LONG).show()
     }
 
     // Function Log Simple Info without message params
@@ -125,8 +119,8 @@ object FrogoLog : IFrogoLog {
 
     // Function Log Info
     override fun i(msg: String?, context: Context) {
-        Log.i(tag(), "${lineNumber()}: $msg")
-        Toast.makeText(context, "${lineNumber()}: $msg", Toast.LENGTH_LONG).show()
+        Log.i(tag(), "$msg")
+        Toast.makeText(context, "$msg", Toast.LENGTH_LONG).show()
     }
 
     // Function Log Simple Warn without message params
@@ -137,20 +131,14 @@ object FrogoLog : IFrogoLog {
 
     // Function Log Warn
     override fun w(msg: String?, context: Context) {
-        Log.w(tag(), "${lineNumber()}: $msg")
-        Toast.makeText(context, "${lineNumber()}: $msg", Toast.LENGTH_LONG).show()
+        Log.w(tag(), "$msg")
+        Toast.makeText(context, "$msg", Toast.LENGTH_LONG).show()
     }
 
     // Function Log Warn
     override fun w(e: Exception?, context: Context) {
-        Log.w(tag(), "${lineNumber()}: ${e?.localizedMessage}")
-        Toast.makeText(context, "${lineNumber()}: ${e?.localizedMessage}", Toast.LENGTH_LONG).show()
-    }
-
-    // Function Log Warn
-    override fun w(e: Throwable?, context: Context) {
-        Log.w(tag(), "${lineNumber()}: ${e?.localizedMessage}")
-        Toast.makeText(context, "${lineNumber()}: ${e?.localizedMessage}", Toast.LENGTH_LONG).show()
+        Log.w(tag(), "${e?.localizedMessage}")
+        Toast.makeText(context, "${e?.localizedMessage}", Toast.LENGTH_LONG).show()
     }
 
     // Function Log Simple Error without message params
@@ -159,10 +147,16 @@ object FrogoLog : IFrogoLog {
         Toast.makeText(context, SIMPLE_MESSSAGE, Toast.LENGTH_LONG).show()
     }
 
+    // Function Log Warn
+    override fun w(e: Throwable?, context: Context) {
+        Log.w(tag(), "${e?.localizedMessage}")
+        Toast.makeText(context, "${e?.localizedMessage}", Toast.LENGTH_LONG).show()
+    }
+
     // Function Log Error
     override fun e(msg: String?, context: Context) {
-        Log.e(tag(), "${lineNumber()}: $msg")
-        Toast.makeText(context, "${lineNumber()}: $msg", Toast.LENGTH_LONG).show()
+        Log.e(tag(), "$msg")
+        Toast.makeText(context, "$msg", Toast.LENGTH_LONG).show()
     }
 
 }
