+++ frogosdk/src/main/java/com/frogobox/sdk/log/LogConstant.kt
@@ -1,4 +1,4 @@
-package com.frogobox.log
+package com.frogobox.sdk.log
 
 
 /**
